+++ bioodo/rseqc.py
@@ -2,7 +2,7 @@
 import re
 import pandas as pd
 import bioodo
-from bioodo import resource, annotate_by_uri, DataFrame, utils
+from bioodo import resource, annotate_by_uri, pivot, DataFrame, utils
 import logging
 
 
@@ -12,15 +12,18 @@ config = bioodo.__RESOURCE_CONFIG__['rseqc']
 
 @resource.register(config['read_distribution']['pattern'],
                    priority=config['read_distribution']['priority'])
+@pivot
 @annotate_by_uri
 def resource_read_distribution(uri, **kwargs):
-    df = pd.read_table(uri, skiprows=list(range(0, 4)) + [15], sep="[ ]+", engine="python")
+    df = pd.read_table(uri, skiprows=list(range(0, 4)) + [15],
+                       sep="[ ]+", engine="python")
     df = df.set_index("Group")
     return df
 
 
 @resource.register(config['data_frame']['pattern'],
                    priority=config['data_frame']['priority'])
+@pivot
 @annotate_by_uri
 def resource_data_frame(uri, **kwargs):
     df = pd.read_table(uri, header=0)
@@ -29,6 +32,7 @@ def resource_data_frame(uri, **kwargs):
 
 @resource.register(config['clipping_profile']['pattern'],
                    priority=config['clipping_profile']['priority'])
+@pivot
 @annotate_by_uri
 def resource_clipping_profile(uri, **kwargs):
     pe = False
@@ -39,10 +43,12 @@ def resource_clipping_profile(uri, **kwargs):
         pe = True
     if pe:
         sections = re.split("\nRead-2:\n", data)
-        df1 = DataFrame.from_records([x.split("\t") for x in sections[0].split("\n")[2:]])
+        df1 = DataFrame.from_records(
+            [x.split("\t") for x in sections[0].split("\n")[2:]])
         df1.columns = header.split("\t")
         df1["Read"] = 1
-        df2 = DataFrame.from_records([x.split("\t") for x in sections[1].split("\n")])
+        df2 = DataFrame.from_records(
+            [x.split("\t") for x in sections[1].split("\n")])
         df2.columns = header.split("\t")
         df2["Read"] = 2
         df = pd.concat([df1, df2])
@@ -54,6 +60,7 @@ def resource_clipping_profile(uri, **kwargs):
 
 @resource.register(config['xls']['pattern'],
                    priority=config['xls']['priority'])
+@pivot
 @annotate_by_uri
 def resource_xls(uri, **kwargs):
     df = pd.read_table(uri, header=0)