+++ HISTORY.rst
@@ -2,6 +2,20 @@
 History
 =======
 
+0.1.0 (2017-10-23)
+------------------
+
+
+* Add multiple pattern to picard insert metrics (#15)
+* Add support for wide table output (#12)
+* Add regex annotation to annotate_by_uri (#13)
+* Add parser for samtools idxstats (#11)
+
+
+
+Alpha releases
+---------------
+
 0.1.0-alpha.1 (2017-09-25)
 --------------------------